:truck: Chore: Move receipts manager to it's own package
