+++ packages/receipt-manager/.depcheckrc
@@ -1,0 +1,4 @@
+ignores: [
+  "@tevm/common"
+]
+skip-missing: true
