+++ packages/receipt-manager/.gitignore
@@ -1,0 +1,50 @@
+.env
+
+cache
+forge-artifacts
+broadcast
+
+# compiled output
+dist
+types
+packages/*/dist
+tmp
+/out-tsc
+**/tsconfig.tsbuildinfo
+
+# dependencies
+node_modules
+
+# IDEs and editors
+/.idea
+.project
+.classpath
+.c9/
+*.launch
+.settings/
+*.sublime-workspace
+
+**/lcov.info
+
+# IDE - VSCode
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+
+# misc
+/.sass-cache
+/connect.lock
+/coverage
+/libpeerconnection.log
+npm-debug.log
+testem.log
+/typings
+
+# System Files
+.DS_Store
+Thumbs.db
+
+# My personal files
+.zshrc
