+++ packages/receipt-manager/docs/API.md
@@ -1,0 +1,36 @@
+**@tevm/blockchain** ∙ [README](README.md) ∙ [API](API.md)
+
+***
+
+# API
+
+## Exports
+
+### Classes
+
+- [ReceiptsManager](classes/ReceiptsManager.md)
+
+### Interfaces
+
+- [BaseTxReceipt](interfaces/BaseTxReceipt.md)
+- [EIP4844BlobTxReceipt](interfaces/EIP4844BlobTxReceipt.md)
+- [MetaDBManagerOptions](interfaces/MetaDBManagerOptions.md)
+- [PostByzantiumTxReceipt](interfaces/PostByzantiumTxReceipt.md)
+- [PreByzantiumTxReceipt](interfaces/PreByzantiumTxReceipt.md)
+
+### Type Aliases
+
+- [Chain](type-aliases/Chain.md)
+- [ChainOptions](type-aliases/ChainOptions.md)
+- [DbType](type-aliases/DbType.md)
+- [MapDb](type-aliases/MapDb.md)
+- [TxReceipt](type-aliases/TxReceipt.md)
+- [TxReceiptWithType](type-aliases/TxReceiptWithType.md)
+
+### Functions
+
+- [createChain](functions/createChain.md)
+- [createMapDb](functions/createMapDb.md)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
