+++ packages/receipt-manager/docs/README.md
@@ -1,0 +1,42 @@
+**@tevm/blockchain** ∙ [README](README.md) ∙ [API](API.md)
+
+***
+
+<p align="center">
+  <a href="https://tevm.sh/">
+    <picture>
+      <source media="(prefers-color-scheme: dark)" srcset="https://user-images.githubusercontent.com/35039927/218812217-92f0f784-cb85-43b9-9ca6-e2b9effd9eb2.png">
+      <img alt="wagmi logo" src="https://user-images.githubusercontent.com/35039927/218812217-92f0f784-cb85-43b9-9ca6-e2b9effd9eb2.png" width="auto" height="300">
+    </picture>
+  </a>
+</p>
+
+<p align="center">
+  Execute solidity scripts in browser
+<p>
+
+[![CI](https://github.com/evmts/tevm-monorepo/actions/workflows/e2e.yml/badge.svg)](https://github.com/evmts/tevm-monorepo/actions/workflows/e2e.yml)
+[![CI](https://github.com/evmts/tevm-monorepo/actions/workflows/unit.yml/badge.svg)](https://github.com/evmts/tevm-monorepo/actions/workflows/unit.yml)
+<a href="https://www.npmjs.com/package/@tevm/blockchain" target="\_parent">
+<img alt="" src="https://img.shields.io/npm/dm/@tevm/blockchain.svg" />
+</a>
+<a href="https://bundlephobia.com/package/@tevm/blockchain@latest" target="\_parent">
+<img alt="" src="https://badgen.net/bundlephobia/minzip/@tevm/blockchain" />
+</a>
+
+# @tevm/blockchain
+
+Tevm vm blockchain package for holding tevm blockchain.
+
+Some files are adapted from [ethereumjs](https://github.com/ethereumjs/ethereumjs-monorepo) and are thus licensed under [mozilla public license 2.0](https://github.com/ethereumjs/ethereumjs-monorepo/blob/master/packages/statemanager/LICENSE)
+
+## Visit [Docs](https://tevm.sh/) for docs, guides, API and more!
+
+## See [Tevm Beta project board](https://github.com/orgs/tevm/projects/1) for progress on the upcoming beta release
+
+## License 📄
+
+<a href="./LICENSE"><img src="https://user-images.githubusercontent.com/35039927/231030761-66f5ce58-a4e9-4695-b1fe-255b1bceac92.png" width="200" /></a>
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
