+++ packages/receipt-manager/docs/functions/createChain.md
@@ -1,0 +1,22 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > createChain
+
+# Function: createChain()
+
+> **createChain**(`options`): `Promise`\<[`Chain`](../type-aliases/Chain.md)\>
+
+## Parameters
+
+▪ **options**: [`ChainOptions`](../type-aliases/ChainOptions.md)
+
+## Returns
+
+## Source
+
+[createChain.js:15](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/createChain.js#L15)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
