+++ packages/receipt-manager/docs/functions/createMapDb.md
@@ -1,0 +1,24 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > createMapDb
+
+# Function: createMapDb()
+
+> **createMapDb**(`options`): [`MapDb`](../type-aliases/MapDb.md)
+
+Creates a {MapDb} which uses an in memory map as it's underlying data structure
+
+## Parameters
+
+▪ **options**: [`MetaDBManagerOptions`](../interfaces/MetaDBManagerOptions.md)
+
+## Returns
+
+## Source
+
+[createMapDb.js:24](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/createMapDb.js#L24)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
