+++ packages/receipt-manager/docs/interfaces/MetaDBManagerOptions.md
@@ -1,0 +1,20 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > MetaDBManagerOptions
+
+# Interface: MetaDBManagerOptions
+
+## Properties
+
+### cache
+
+> **cache**: `Map`\<\`0x${string}\`, `Uint8Array`\>
+
+#### Source
+
+[MapDb.ts:14](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/MapDb.ts#L14)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
