+++ packages/receipt-manager/docs/interfaces/PreByzantiumTxReceipt.md
@@ -1,0 +1,77 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > PreByzantiumTxReceipt
+
+# Interface: PreByzantiumTxReceipt
+
+Pre-Byzantium receipt type with a field
+for the intermediary state root
+
+## Extends
+
+- [`BaseTxReceipt`](BaseTxReceipt.md)
+
+## Properties
+
+### bitvector
+
+> **bitvector**: `Uint8Array`
+
+Bloom bitvector
+
+#### Inherited from
+
+[`BaseTxReceipt`](BaseTxReceipt.md).[`bitvector`](BaseTxReceipt.md#bitvector)
+
+#### Source
+
+[RecieptManager.ts:25](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/RecieptManager.ts#L25)
+
+***
+
+### cumulativeBlockGasUsed
+
+> **cumulativeBlockGasUsed**: `bigint`
+
+Cumulative gas used in the block including this tx
+
+#### Inherited from
+
+[`BaseTxReceipt`](BaseTxReceipt.md).[`cumulativeBlockGasUsed`](BaseTxReceipt.md#cumulativeblockgasused)
+
+#### Source
+
+[RecieptManager.ts:21](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/RecieptManager.ts#L21)
+
+***
+
+### logs
+
+> **logs**: `Log`[]
+
+Logs emitted
+
+#### Inherited from
+
+[`BaseTxReceipt`](BaseTxReceipt.md).[`logs`](BaseTxReceipt.md#logs)
+
+#### Source
+
+[RecieptManager.ts:29](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/RecieptManager.ts#L29)
+
+***
+
+### stateRoot
+
+> **stateRoot**: `Uint8Array`
+
+Intermediary state root
+
+#### Source
+
+[RecieptManager.ts:49](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/RecieptManager.ts#L49)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
