+++ packages/receipt-manager/docs/type-aliases/DbType.md
@@ -1,0 +1,16 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > DbType
+
+# Type alias: DbType
+
+> **DbType**: `"Receipts"` \| `"TxHash"` \| `"SkeletonBlock"` \| `"SkeletonBlockHashToNumber"` \| `"SkeletonStatus"` \| `"SkeletonUnfinalizedBlockByHash"` \| `"Preimage"`
+
+## Source
+
+[MapDb.ts:4](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/MapDb.ts#L4)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
