+++ packages/receipt-manager/docs/type-aliases/MapDb.md
@@ -1,0 +1,46 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > MapDb
+
+# Type alias: MapDb
+
+> **MapDb**: `object`
+
+Helper class to access the metaDB with methods `put`, `get`, and `delete`
+
+## Type declaration
+
+### delete()
+
+#### Parameters
+
+▪ **type**: [`DbType`](DbType.md)
+
+▪ **hash**: `Uint8Array`
+
+### get()
+
+#### Parameters
+
+▪ **type**: [`DbType`](DbType.md)
+
+▪ **hash**: `Uint8Array`
+
+### put()
+
+#### Parameters
+
+▪ **type**: [`DbType`](DbType.md)
+
+▪ **hash**: `Uint8Array`
+
+▪ **value**: `Uint8Array`
+
+## Source
+
+[MapDb.ts:20](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/MapDb.ts#L20)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
