+++ packages/receipt-manager/docs/type-aliases/TxReceiptWithType.md
@@ -1,0 +1,19 @@
+**@tevm/blockchain** ∙ [README](../README.md) ∙ [API](../API.md)
+
+***
+
+[API](../API.md) > TxReceiptWithType
+
+# Type alias: TxReceiptWithType
+
+> **TxReceiptWithType**: `PreByzantiumTxReceiptWithType` \| `PostByzantiumTxReceiptWithType`
+
+TxReceiptWithType extends TxReceipt to provide:
+ - txType: byte prefix for serializing typed tx receipts
+
+## Source
+
+[RecieptManager.ts:72](https://github.com/evmts/tevm-monorepo/blob/main/packages/blockchain/src/RecieptManager.ts#L72)
+
+***
+Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
