+++ packages/receipt-manager/package.json
@@ -1,0 +1,84 @@
+{
+	"name": "@tevm/receipt-manager",
+	"version": "1.1.0-next.47",
+	"private": false,
+	"description": "A custom implementation of ethereumjs receipt-manager",
+	"keywords": [
+		"solidity",
+		"ethereumjs",
+		"typescript",
+		"web3",
+		"blockchain"
+	],
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/evmts/tevm-monorepo.git",
+		"directory": "packages/blockchain"
+	},
+	"license": "MIT",
+	"contributors": [
+		"Will Cory <willcory10@gmail.com>"
+	],
+	"type": "module",
+	"exports": {
+		"./package.json": "./package.json",
+		".": {
+			"import": {
+				"types": "./types/index.d.ts",
+				"default": "./dist/index.js"
+			},
+			"require": {
+				"types": "./dist/index.d.cts",
+				"default": "./dist/index.cjs"
+			}
+		}
+	},
+	"main": "dist/index.cjs",
+	"module": "dist/index.js",
+	"types": "types/index.d.ts",
+	"files": [
+		"dist",
+		"types",
+		"src",
+		"!src/**/*.spec.ts"
+	],
+	"scripts": {
+		"all": "pnpm i  && bun run build && bun lint && bun format && bun test:run && bun generate:docs",
+		"build": "nx run-many --targets=build:dist,build:types --projects=@tevm/receipt-manager",
+		"build:dist": "tsup",
+		"build:types": "tsup --dts-only && tsc --emitDeclarationOnly --declaration",
+		"clean": "rm -rf node_modules && rm -rf artifacts && rm -rf dist && rm -rf cache",
+		"format": "biome format . --write",
+		"format:check": "biome format .",
+		"generate:docs": "typedoc",
+		"lint": "biome check . --apply-unsafe",
+		"lint:check": "biome check . --verbose",
+		"lint:deps": "bunx depcheck",
+		"lint:package": "bunx publint --strict && attw --pack",
+		"package:up": "pnpm up --latest",
+		"test": "bun test --watch",
+		"test:coverage": "bun test --coverage",
+		"test:run": "bun test",
+		"typecheck": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@ethereumjs/blockchain": "^7.2.0",
+		"@tevm/block": "workspace:^",
+		"@tevm/blockchain": "workspace:1.1.0-next.47",
+		"@tevm/jsonrpc": "workspace:^",
+		"@tevm/logger": "workspace:^",
+		"@tevm/rlp": "workspace:^",
+		"@tevm/trie": "workspace:^",
+		"@tevm/tx": "workspace:^",
+		"viem": "^2.9.28"
+	},
+	"devDependencies": {
+		"@tevm/common": "workspace:^",
+		"@tevm/tsconfig": "workspace:^",
+		"@tevm/tsupconfig": "workspace:^",
+		"@tevm/utils": "workspace:^"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
