+++ packages/receipt-manager/src/index.ts
@@ -1,4 +1,3 @@
-export { createChain } from './createChain.js'
 export {
   ReceiptsManager,
   type BaseTxReceipt,
@@ -8,5 +7,9 @@ export {
   type TxReceipt,
   type TxReceiptWithType,
 } from './RecieptManager.js'
-export type { Chain } from './Chain.js'
-export type { ChainOptions } from './ChainOptions.js'
+export { createMapDb } from './createMapDb.js'
+export type {
+  MapDb,
+  MetaDBManagerOptions,
+  DbType,
+} from './MapDb.js'
