+++ packages/receipt-manager/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "@tevm/tsconfig/base.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "types",
+		"skipLibCheck": true,
+		"types": ["bun-types"]
+	},
+	"include": ["src"]
+}
