+++ packages/receipt-manager/tsup.config.ts
@@ -1,0 +1,4 @@
+import { createTsUpOptions } from '@tevm/tsupconfig'
+export default createTsUpOptions({
+	entry: ['./src/index.ts'],
+})
